+++ vite-project/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Photo Fetcher SNET</title>
-    <link rel="stylesheet" href="./styles.css" />
+    <link rel="stylesheet" href="src/style.css" />
   </head>
   <body>
     <main class="page">
